requirements.txt
python-telegram-bot==21.0
python-dateutil==2.8.2
pytz==2024.1
imagehash==4.3.1
Pillow==10.2.0
asyncpg==0.29.0
